COMP: fixed warnings

+++ VTK/Widgets/vtkSeedRepresentation.cxx
@@ -206,7 +206,7 @@ void vtkSeedRepresentation::RemoveHandle( int n )
     return;
     }
 
-  if ( this->Handles->size() <= n )
+  if (static_cast<int>(this->Handles->size()) <= n )
     {
     return;
     }
